+++ src/dinostomp/__init__.py
@@ -9,7 +9,7 @@ from dinostomp.spec import (
     validate_obj,
 )
 
-__version__ = "0.39.1"
+__version__ = "0.40.0"
 
 __all__ = [
     "SCHEMA_NAMES",